refactor(db): close database driver seam with unified DatabaseClient
Replace PgliteDatabase|PostgresJsDatabase union across 25 files with a single DatabaseClient seam in src/lib/server/db/db.ts.

- New src/lib/server/db/db.ts exports unified DatabaseClient, getDbClient, and getQueryRows

- Hooks.server.ts now injects db into locals so routes never lazy-import the singleton

- Deleted src/lib/server/utils/db-helpers.ts (absorbed into seam)

- Removed inline getDbClient() copies from 15 route files

- Removed db as {transaction} casts in v1 ingestion routes

- All type checks, lint, knip, unit + integration tests passing
