+++ src/app.d.ts
@@ -7,7 +7,7 @@ declare global {
       user?: import('./lib/server/auth').User;
       session?: import('./lib/server/auth').Session;
       // Optional db client for testing (dependency injection)
-      db?: import('drizzle-orm/pglite').PgliteDatabase<typeof import('./lib/server/db/schema')>;
+      db?: import('./lib/server/db/db').DatabaseClient;
     }
     // interface PageData {}
     // interface PageState {}
