+++ src/lib/server/db/index.ts
@@ -8,3 +8,5 @@ if (!env.DATABASE_URL) throw new Error('DATABASE_URL is not set');
 const client = postgres(env.DATABASE_URL);
 
 export const db = drizzle(client, { schema });
+
+export type { DatabaseClient } from './db';
