+++ src/routes/api/projects/[id]/+server.ts
@@ -1,8 +1,6 @@
 import { json } from '@sveltejs/kit';
 import { and, count, eq, ne } from 'drizzle-orm';
-import type { PgliteDatabase } from 'drizzle-orm/pglite';
-import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
-import type * as schema from '$lib/server/db/schema';
+import { getDbClient } from '$lib/server/db/db';
 import { log, project } from '$lib/server/db/schema';
 import { invalidateApiKeyCache } from '$lib/server/utils/api-key';
 import { requireJsonContentType } from '$lib/server/utils/content-type';
@@ -10,20 +8,6 @@ import { checkCsrfOrigin } from '$lib/server/utils/csrf';
 import { isErrorResponse, requireProjectOwnership } from '$lib/server/utils/project-guard';
 import { projectUpdatePayloadSchema } from '$lib/shared/schemas/project';
 import type { RequestEvent } from './$types';
-
-type DatabaseClient = PostgresJsDatabase<typeof schema> | PgliteDatabase<typeof schema>;
-
-/**
- * Helper to get database client from locals or production db
- * Supports test injection via locals.db
- */
-async function getDbClient(locals: App.Locals): Promise<DatabaseClient> {
-  if (locals.db) {
-    return locals.db as DatabaseClient;
-  }
-  const { db } = await import('$lib/server/db');
-  return db;
-}
 
 /**
  * GET /api/projects/[id]
